Rename packages in favor of Go's naming convention
